Fix connection pooling for PostgresStorage... Apparently the PostgresQL JDBC driver silently ignores StatementEventListeners.
In order to apply the above fix H2PreparedStatement was moved out into the generic 'database' package as ManagedStatement. Now Postgres applies the same workaround as H2 to get statement pooling working.

Fix another bug in the PostgresStorage related to integrity constraints. Postgres, like H2, does not throw IntegrityConstraintViolationExceptions when they occur, but use standard SQLExceptions with a specific SQLState (really the error code stored here instead of the errorCode variable of the SQLException... wonder... wonder...)

Add a method DatabaseStorage.destroyDatabase() that drops all created tables. Ofcourse with a BIG caveat emptor in the docs! This was needed to get the unit tests running for PostgresStorage (which they now do by the way...)