WARNING: CHANGE OF DATABASE SCHEMA!
Switch DatabaseStorage to use UniqueTimestampGenerator timestamps instead of DATETIME columns. That is, use BIGINT columns with the generated timestamps.

This means that we can also add a UNIQUE index on the mtime column - which again means that we can drop the "id" part of "ORDER BY mtime,id".

The longer term point of this is to allow super-fast paginated iteration using only the mtime as keys for the rows.