WARNING: New database table layout!
Add two methods hasChildren() and hasParents() to Record. This allows the DatabaseStorage to not construct a gajillion ArrayLists for eachRecord in each flush()

Impl of lazy relation lookups more or less finished, now I just need to test it and make H2Storage use the new features. Changes include:
 - Add hasRelations column to summa_records table
 - Add code to make sure that hasRelations is always up to date
 - Make scanRecord() resolve ids lazily if it makes sense AND it is told to do so
 - Impl helper methods to keep relation flag in sync: markHasRelations(List<String> ids) and checkHasRelations(String id)

Fix bugs with invalid parameter list lengths in the SQL queries of updateRecord() and createRecord()