Huge cleanup of the entire Lucene Analyzer/Tokenizer/Filter framework under Summa. This is guaranteed to break a lot of things.
Add a cache to RuleParser which checks if the rules to parse has already been parsed and reuses them if so.

Rename SummaRepeatAnalyzer to SummaFieldSeparatingAnalyzer

Rename SummaSortKeyAnalyzer to SummaSymbolRemovingAnalyzer

Make all Analyzer subclasses implement the Analyzer.reusableTokenStream(String,Reader) method

Make TransliteratorTokenizer only implement the TokenStream.next(Token) and not TokenSteam.next()

Make  TransliteratorTokenizer not allocate new StringBuffers in .next(Token), but simply reset the buffer with StringBuffer.setLength(0)

Add a CharSequenceReader with unit tests that acts as a Reader on top of CharSequence implementors such as StringBuilder and StringBuffer. This allows us to do a lot less object allocations when altering terms in-memory using a StringBuffer and the likes